Revert "fix: change uint32 type to int64 type"
This reverts commit 85a6c171c26a03cdb02a951ebf402309f2ca97ae.

+++ src/zipline/data/continuous_future_reader.py
@@ -265,7 +265,7 @@ class ContinuousFutureMinuteBarReader(SessionBarReader):
             if column != "volume":
                 out = np.full(shape, np.nan)
             else:
-                out = np.zeros(shape, dtype=np.int64)
+                out = np.zeros(shape, dtype=np.uint32)
             for i, asset in enumerate(assets):
                 partitions = partitions_by_asset[asset]
                 for sid, start, end, start_loc, end_loc in partitions:
